Add Samjin multi sensor variation without Diagnostic cluster
